[refactor] 보드 저장부분 엔티티로 이동
