+++ src/main/java/chess/util/BooleanYNConverter.java
@@ -1,0 +1,8 @@
+package chess.util;
+
+public class BooleanYNConverter {
+
+    public static String convertYN(boolean changer) {
+        return changer ? "Y" : "N";
+    }
+}
